Time: 3 ms (8.08%) | Memory: 6.5 MB (11.11%) - LeetSync

+++ 2628-minimize-the-maximum-of-two-arrays/minimize-the-maximum-of-two-arrays.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int minimizeSet(int d1, int d2, int uc1, int uc2) {
-        int s=1,e=INT_MAX;
+        long long s=1,e=INT_MAX;
         long long lcm = (long long)d1*d2/gcd(d1, d2);
         int ans=1;
         while(s<=e){
